+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -441,6 +441,10 @@ class GeckoEngineSession(
         geckoSession.close()
     }
 
+    override fun getBlockedSchemes(): List<String> {
+        return BLOCKED_SCHEMES
+    }
+
     /**
      * NavigationDelegate implementation for forwarding callbacks to observers of the session.
      */